testing/testing_gemini.py
import google.generativeai as genai
from dotenv import load_dotenv
import os
load_dotenv()

genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))
model = genai.GenerativeModel(model_name="gemini-1.5-flash")
response = model.generate_content("give me summary of book atomic habits in under 500 words?")

print(response.text)