requirements.txt
streamlit
google-generativeai
youtube_transcript_api
python-dotenv
pathlib